Rewrite OWNERS docs
Move from design doc to state-of-the-world doc

The OWNERS feature has been implemented for a while now. Let's
document how it actually works

I originally started this with the intent of fleshing out measurable
critera for adding/removing OWNERS, but as I dug into it, I realized I
need more time to explore useful metrics.  Let's find out if this is
enough of a rewrite to call incremental progress